Joe Sources 3.1 ( 2004-05-29 )

+++ tty.c
@@ -27,6 +27,13 @@
 #ifdef HAVE_SYS_PARAM_H
 #include <sys/param.h>
 #endif
+
+#ifdef HAVE_OPENPTY
+#ifdef HAVE_PTY_H
+#include <pty.h>
+#endif
+#endif
+
 #include <errno.h>
 #include <string.h>
 #include <stdlib.h>
@@ -257,7 +264,6 @@ extern int dostaupd;
 static RETSIGTYPE dotick(int unused)
 {
 	ticked = 1;
-	dostaupd = 1;
 }
 
 void tickoff(void)
@@ -298,7 +304,6 @@ void ttopnn(void)
 #ifdef SIGWINCH
 			joe_set_signal(SIGWINCH, winchd);
 #endif
-			tickon();
 		}
 	}
 
@@ -497,7 +502,6 @@ int ttflsh(void)
 			while (!yep)
 				pauseit();
 			unmaskit();
-			tickon();
 		} else
 			joe_write(fileno(termout), obuf, obufp);
 
@@ -557,11 +561,23 @@ int ttflsh(void)
 
 void mpxdied(MPX *m);
 
+long last_time;
+
 int ttgetc(void)
 {
 	int stat;
+	long new_time;
+
+
+	tickon();
 
       loop:
+	new_time = time(NULL);
+	if (new_time != last_time) {
+		last_time = new_time;
+		dostaupd = 1;
+		ticked = 1;
+	}
 	ttflsh();
 	while (winched) {
 		winched = 0;
@@ -598,9 +614,12 @@ int ttgetc(void)
 				mpxdied(pack.who);
 			goto loop;
 		} else {
-			if (accept != NO_MORE_DATA)
+			if (accept != NO_MORE_DATA) {
+				tickoff();
 				return accept;
+			}
 			else {
+				tickoff();
 				ttsig(0);
 				return 0;
 			}
@@ -616,6 +635,7 @@ int ttgetc(void)
 				ttsig(0);
 		}
 	}
+	tickoff();
 	return havec;
 }
 
@@ -665,8 +685,10 @@ void ttshell(unsigned char *cmd)
 	int x, omode = ttymode;
 	unsigned char *s = (unsigned char *)getenv("SHELL");
 
-	if (!s)
-		return;
+	if (!s) {
+		s = US "/bin/sh";
+		/* return; */
+	}
 	ttclsn();
 	if ((x = fork()) != 0) {
 		if (x != -1)
@@ -685,40 +707,11 @@ void ttshell(unsigned char *cmd)
 	}
 }
 
-void ttsusp(void)
-{
-	int omode;
+/* Create keyboard task */
 
-	tickoff();
-#ifdef SIGTSTP
-	omode = ttymode;
-	ttclsn();
-	fprintf(stderr, "You have suspended the program.  Type 'fg' to return\n");
-	kill(0, SIGTSTP);
-	if (ackkbd != -1)
-		kill(kbdpid, SIGCONT);
-	if (omode)
-		ttopnn();
-#else
-	ttshell(NULL);
-#endif
-	tickon();
-}
-
-/* Stuff for asynchronous I/O multiplexing.  We do not use streams or select() because joe
-   needs to work on versions of UNIX which predate these calls.  Instead, when there is
-   multiple async sources, we use helper processes which packetize data from the sources.  A
-   header on each packet indicates the source.  There is no guarentee that packets getting
-   written to the same pipe don't get interleaved, but you can reasonable rely on it with
-   small packets. */
-
-static void mpxstart(void)
+static void mpxresume(void)
 {
 	int fds[2];
-
-	pipe(fds);
-	mpxfd = fds[0];
-	mpxsfd = fds[1];
 	pipe(fds);
 	accept = NO_MORE_DATA;
 	have = 0;
@@ -743,12 +736,69 @@ static void mpxstart(void)
 	ackkbd = fds[1];
 }
 
+/* Kill keyboard task */
+
+static void mpxsusp(void)
+{
+	if (ackkbd!=-1) {
+		kill(kbdpid, 9);
+		while (wait(NULL) < 0 && errno == EINTR)
+			/* do nothing */;
+		close(ackkbd);
+	}
+}
+
+/* We used to leave the keyboard copy task around during suspend, but
+   Cygwin gets confused when two processes are waiting for input and you
+   change the tty from raw to cooked (on the call to ttopnn()): the keyboard
+   process was stuck in cooked until he got a carriage return- then he
+   switched back to raw (he's supposed to switch to raw without waiting for
+   the end of line). Probably this should be done for ttshell() as well. */
+
+void ttsusp(void)
+{
+	int omode;
+
+#ifdef SIGTSTP
+	omode = ttymode;
+	mpxsusp();
+	ttclsn();
+	fprintf(stderr, "You have suspended the program.  Type 'fg' to return\n");
+	kill(0, SIGTSTP);
+#ifdef junk
+	/* Hmmm... this should not have been necessary */
+	if (ackkbd != -1)
+		kill(kbdpid, SIGCONT);
+#endif
+	if (omode)
+		ttopnn();
+	if (ackkbd!= -1)
+		mpxresume();
+#else
+	ttshell(NULL);
+#endif
+}
+
+/* Stuff for asynchronous I/O multiplexing.  We do not use streams or
+   select() because joe needs to work on versions of UNIX which predate
+   these calls.  Instead, when there is multiple async sources, we use
+   helper processes which packetize data from the sources.  A header on each
+   packet indicates the source.  There is no guarentee that packets getting
+   written to the same pipe don't get interleaved, but you can reasonable
+   rely on it with small packets. */
+
+static void mpxstart(void)
+{
+	int fds[2];
+	pipe(fds);
+	mpxfd = fds[0];
+	mpxsfd = fds[1];
+	mpxresume();
+}
+
 static void mpxend(void)
 {
-	kill(kbdpid, 9);
-	while (wait(NULL) < 0 && errno == EINTR)
-		/* do nothing */;
-	close(ackkbd);
+	mpxsusp();
 	ackkbd = -1;
 	close(mpxfd);
 	close(mpxsfd);
@@ -772,11 +822,11 @@ static void mpxend(void)
 
 /* Newer sgi machines can do it the __svr4__ way, but old ones can't */
 
-extern unsigned char *_getpty();
+extern char *_getpty();
 
 static unsigned char *getpty(int *ptyfd)
 {
-	return _getpty(ptyfd, O_RDWR, 0600, 0);
+	return (unsigned char *)_getpty(ptyfd, O_RDWR, 0600, 0);
 }
 
 #else
@@ -784,7 +834,7 @@ static unsigned char *getpty(int *ptyfd)
 
 /* Strange streams way */
 
-extern unsigned char *ptsname();
+extern char *ptsname();
 
 static unsigned char *getpty(int *ptyfd)
 {
@@ -794,11 +844,27 @@ static unsigned char *getpty(int *ptyfd)
 	*ptyfd = fdm = open("/dev/ptmx", O_RDWR);
 	grantpt(fdm);
 	unlockpt(fdm);
-	return ptsname(fdm);
+	return (unsigned char *)ptsname(fdm);
 }
 
 #else
+#ifdef HAVE_OPENPTY
 
+/* BSD function, present in libc5 and glibc2 */
+
+static unsigned char *getpty(int *ptyfd)
+{
+	int fdm;
+	static unsigned char name[32];
+	int ttyfd;
+
+        if (openpty(ptyfd, &ttyfd, name, NULL, NULL) == 0)
+           return(name);
+        else
+	   return (NULL);
+}
+
+#else
 /* The normal way: for each possible pty/tty pair, try to open the pty and
  * then the corresponding tty.  If both could be opened, close them both and
  * then re-open the pty.  If that succeeded, return with the opened pty and the
@@ -862,11 +928,16 @@ static unsigned char *getpty(int *ptyfd)
 
 #endif
 #endif
+#endif
 
+/* Shell dies signal handler.  Puts pty in non-block mode so
+ * that read returns with <1 when all data from process has
+ * been read. */
 int dead = 0;
-
+int death_fd;
 static RETSIGTYPE death(int unused)
 {
+	fcntl(death_fd,F_SETFL,O_NDELAY);
 	wait(NULL);
 	dead = 1;
 }
@@ -928,6 +999,10 @@ MPX *mpxmk(int *ptyfd, unsigned char *cmd, unsigned char **args, void (*func) (/
 	if (x==NPROC)
 		return NULL;
 
+	/* Fixes cygwin console bug: if you fork() with inverse video he assumes you want
+	 * ESC [ 0 m to keep it in inverse video from then on. */
+	set_attr(maint->t,0);
+
 	/* Flush output */
 	ttflsh();
 
@@ -964,6 +1039,7 @@ MPX *mpxmk(int *ptyfd, unsigned char *cmd, unsigned char **args, void (*func) (/
 
 		/* Flag which indicates child died */
 		dead = 0;
+		death_fd = *ptyfd;
 		joe_set_signal(SIGCHLD, death);
 
 		if (!(pid = fork())) {
@@ -977,6 +1053,7 @@ MPX *mpxmk(int *ptyfd, unsigned char *cmd, unsigned char **args, void (*func) (/
 			   controlling tty (session leader) and starting a new
 			   session.  This is the most non-portable part of UNIX- second
 			   only to pty/tty pair creation. */
+#ifndef HAVE_LOGIN_TTY
 
 #ifdef TIOCNOTTY
 			x = open("/dev/tty", O_RDWR);
@@ -990,7 +1067,7 @@ MPX *mpxmk(int *ptyfd, unsigned char *cmd, unsigned char **args, void (*func) (/
 			setpgrp();
 #endif
 
-
+#endif
 			/* Close all fds */
 			for (x = 0; x != 32; ++x)
 				close(x);	/* Yes, this is quite a kludge... all in the
@@ -1000,6 +1077,10 @@ MPX *mpxmk(int *ptyfd, unsigned char *cmd, unsigned char **args, void (*func) (/
 			if ((x = open((char *)name, O_RDWR)) != -1) {	/* Standard input */
 				unsigned char **env = newenv(mainenv, US "TERM=");
 
+#ifdef HAVE_LOGIN_TTY
+				login_tty(x);
+
+#else
 				/* This tells the fd that it's a tty (I think) */
 #ifdef __svr4__
 				ioctl(x, I_PUSH, "ptem");
@@ -1011,6 +1092,7 @@ MPX *mpxmk(int *ptyfd, unsigned char *cmd, unsigned char **args, void (*func) (/
 				dup(x);	/* Standard output, standard error */
 				/* (yes, stdin, stdout, and stderr must all be open for reading and
 				 * writing.  On some systems the shell assumes this */
+#endif
 
 				/* We could probably have a special TTY set-up for JOE, but for now
 				 * we'll just use the TTY setup for the TTY was was run on */
@@ -1030,7 +1112,7 @@ MPX *mpxmk(int *ptyfd, unsigned char *cmd, unsigned char **args, void (*func) (/
 				execve((char *)cmd, (char **)args, (char **)env);
 
 				/* If shell didn't execute */
-				snprintf((char *)buf,80,"Couldn't execute shell '%s'\n",cmd);
+				joe_snprintf_1((char *)buf,sizeof(buf),"Couldn't execute shell '%s'\n",cmd);
 				write(0,(char *)buf,strlen((char *)buf));
 				sleep(1);
 			}
@@ -1041,25 +1123,37 @@ MPX *mpxmk(int *ptyfd, unsigned char *cmd, unsigned char **args, void (*func) (/
 		/* Tell JOE PID of shell */
 		joe_write(comm[1], &pid, sizeof(pid));
 
+		/* sigpipe should be ignored here. */
+
 		/* This process copies data from shell to JOE until EOF.  It creates a packet
 		   for each data */
+
+
+		/* We don't really get EOF from a pty- it would just wait forever
+		   until someone else writes to the tty.  So: when the shell
+		   dies, the child died signal handler death() puts pty in non-block
+		   mode.  This allows us to read any remaining data- then
+		   read returns 0 and we know we're done. */
+
 	      loop:
 		pack.who = m;
 		pack.ch = 0;
-		if (dead)
-			pack.size = 0;
-		else {
-			pack.size = read(*ptyfd, pack.data, 1024);
-			/* On SUNOS 5.8, the very first read from the pty returns 0 for some reason */
-			if (!pack.size)
-				pack.size = read(*ptyfd, pack.data, 1024);
-		}
+
+		/* Read data from process */
+		pack.size = joe_read(*ptyfd, pack.data, 1024);
+
+		/* On SUNOS 5.8, the very first read from the pty returns 0 for some reason */
+		if (!pack.size)
+			pack.size = joe_read(*ptyfd, pack.data, 1024);
+
 		if (pack.size > 0) {
+			/* Send data to JOE, wait for ack */
 			joe_write(mpxsfd, &pack, sizeof(struct packet) - 1024 + pack.size);
 
 			joe_read(fds[0], &pack, 1);
 			goto loop;
 		} else {
+			/* Shell died: return */
 			pack.ch = NO_MORE_DATA;
 			pack.size = 0;
 			joe_write(mpxsfd, &pack, sizeof(struct packet) - 1024);
